Try to fix functionality
