+++ src/components/AreaCard/AreaCard.js
@@ -31,7 +31,7 @@ const AreaCard = ({ areaInfo, fetchListings }) => {
       <p>({areaInfo.nickname})</p>
       <p>{areaInfo.about}</p>
       <Link to={`/areas/${areaInfo.id}/listings`}>
-      <button id={areaInfo.id} onClick={(e) => fetchListings(e.target.id)}>LISTINGS</button>
+        <button id={areaInfo.id}>LISTINGS</button>
       </Link>
     </div>
   )
