+++ src/components/Header/Header.css
@@ -12,12 +12,12 @@ header {
   display: block;
   font-size: 45px;
   text-transform: uppercase;
-  padding: 0 6%;
+  padding: 0 10px;
 }
 
 .logo-left-btns, .right-btn {
   display: flex;
-  justify-content: center;
+  justify-content: flex-start;
   align-items: center;
 }
